fix: address CodeRabbit review on GFS hourly retention
- reject fractional GFS values with .int() in both retention schemas
- align onboarding default (hourly 0) with daily schedule
- replace per-bucket .find() scans with single-pass bucket tracking
- enforce min 1 for daily tier in the GFS form

+++ src/features/database/schemas/retention-policy.schema.ts
@@ -1,11 +1,11 @@
 import {z} from "zod";
 
 const GFSSettingsSchema = z.object({
-    hourly: z.number().min(0).max(168),
-    daily: z.number().min(1).max(31),
-    weekly: z.number().min(0).max(52),
-    monthly: z.number().min(0).max(120),
-    yearly: z.number().min(0).max(50),
+    hourly: z.number().int().min(0).max(168),
+    daily: z.number().int().min(1).max(31),
+    weekly: z.number().int().min(0).max(52),
+    monthly: z.number().int().min(0).max(120),
+    yearly: z.number().int().min(0).max(50),
 });
 
 export const RetentionSettingsSchema = z.object({
